% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sliding_window_file.R
\name{apply_sliding_window}
\alias{apply_sliding_window}
\title{Apply sliding window}
\usage{
apply_sliding_window(window_function, data_frame, window_size, threshold = 0)
}
\arguments{
\item{window_function}{the function which will be iteratively applied across each window}

\item{data_frame}{the raw data}

\item{window_size}{the size (length) of the window}
}
\value{
Iterates over time series data and applies the windows functions for every data point
}
\description{
Apply sliding window
}
