% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sliding_window_file.R
\name{binary_filter}
\alias{binary_filter}
\title{Binary filter}
\usage{
binary_filter(window_data, dummy_variable = 0)
}
\arguments{
\item{window_data}{The data to be parsed through the function}
}
\value{
Iterates through a window to returning a +1/-1 depending on whether the value increases or decreases from the beginning to the end of the window
}
\description{
Binary filter
}
