% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{binary_window}
\alias{binary_window}
\title{Binary window}
\usage{
binary_window(window_vec)
}
\arguments{
\item{window_vec}{a vector contaning a window}
}
\value{
a vector containing the binary output
}
\description{
Takes a window and returns a binary value 1 och -1 depending
on direction of acceleration.
}
