% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Caiculation_acceleration function.R
\name{calculate_acceleration}
\alias{calculate_acceleration}
\title{Calculate the Euclidean Norm of Acceleration}
\usage{
calculate_acceleration(file_path, x_col, y_col, z_col)
}
\arguments{
\item{file_path}{Path to the CSV file.}

\item{x_col}{Name or index of the x-axis column.}

\item{y_col}{Name or index of the y-axis column.}

\item{z_col}{Name or index of the z-axis column.}
}
\value{
Data frame with added acceleration vector column.
}
\description{
calculate_acceleration
}
