% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slidingwindowfunction.R
\name{combo_axis}
\alias{combo_axis}
\title{Calculate Euclidean norm}
\usage{
combo_axis(data, x, y, z)
}
\arguments{
\item{data}{dataset}

\item{x}{x axis}

\item{y}{y axis}

\item{z}{z axis}
}
\value{
a data frame
}
\description{
Calculate Euclidean norm
}
