% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sliding_window_file.R
\name{euclidean_norm}
\alias{euclidean_norm}
\title{Calculating euclidean norm}
\usage{
euclidean_norm(data_file_path, dimensions)
}
\arguments{
\item{data_file_path}{path to load data file}

\item{dimensions}{vector containing the dimensions (i.e. # of columns)}
}
\value{
The euclidean norm of the input sequential data
}
\description{
Calculating euclidean norm
}
