% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Iterate_time_series function.R
\name{Iterate_time_series}
\alias{Iterate_time_series}
\title{Iterate Over Time Series}
\usage{
Iterate_time_series(data, window_function, window_size, threshold)
}
\arguments{
\item{data}{Data frame with acceleration data.}

\item{window_function}{Window function to apply.}

\item{window_size}{Size of the window.}

\item{threshold}{Threshold for window function.}
}
\value{
Data frame with added mean and binary columns.
}
\description{
Iterate Over Time Series
}
