% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{load_accel}
\alias{load_accel}
\title{Load Accelerometer File and Compute Acceleration Vector}
\usage{
load_accel(file, x = "x", y = "y", z = "z")
}
\arguments{
\item{file}{file path for a .csv file}

\item{x}{variable to use for x (default = "x")}

\item{y}{variable to use for y (default = "y")}

\item{z}{variable to use for z (default = "z")}
}
\value{
dataframe with combined acceleration vector
}
\description{
Reads CSV and computes Euclidean norm.
}
