% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{plot_accel}
\alias{plot_accel}
\title{plot accel}
\usage{
plot_accel(df, window_col = "window_result")
}
\arguments{
\item{df}{a dataframe}

\item{window_col}{the filtered column}
}
\description{
Creates a plot of the acceleration vector and the filtered vector
}
