% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slidingwindowfunction.R
\name{plot_accelaration}
\alias{plot_accelaration}
\title{Plot acceleration results}
\usage{
plot_accelaration(data, use_mean = TRUE, use_binary = TRUE)
}
\arguments{
\item{data}{data frame}

\item{use_mean}{show mean results}

\item{use_binary}{show binary results}
}
\value{
a data frame
}
\description{
Plot acceleration results
}
