% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot Acceleration and Results function.R
\name{plot_results}
\alias{plot_results}
\title{Plot Acceleration and Results}
\usage{
plot_results(data, show_mean = TRUE, show_binary = TRUE)
}
\arguments{
\item{data}{A data frame containing the columns "time", "acceleration", "mean", and "binary".}

\item{show_mean}{Logical. Whether to show the mean line on the plot. Default is TRUE.}

\item{show_binary}{Logical. Whether to show the binary line on the plot. Default is TRUE.}
}
\value{
A ggplot object displaying the acceleration data and optional mean and binary lines.
}
\description{
Plot Acceleration and Results
}
\details{
This function generates a plot of acceleration data with optional mean and binary lines.
}
\examples{
data <- data.frame(
  time = as.POSIXct(seq(1, 100, by = 1), origin = "1970-01-01"),
  acceleration = rnorm(100),
  mean = rnorm(100),
  binary = sample(c(-1, 0, 1), 100, replace = TRUE)
)
plot_results(data, show_mean = TRUE, show_binary = FALSE)
}
