% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sliding_window_file.R
\name{plot_wrapper}
\alias{plot_wrapper}
\title{ggplot2 wrapper function}
\usage{
plot_wrapper(data, column_names)
}
\arguments{
\item{data}{data frame with a time column and the list of columns in column names}

\item{column_names}{other columns from the data frame to be included in the graph}
}
\value{
returns plots of input columns
}
\description{
ggplot2 wrapper function
}
