% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slidingwindowfunction.R
\name{read_data_frame}
\alias{read_data_frame}
\title{Load a file and convert the time column.}
\usage{
read_data_frame(file_path)
}
\arguments{
\item{file_path}{Path to the CSV file}
}
\value{
a data frame
}
\description{
Load a file and convert the time column.
}
