% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{run_full_binary_pipeline}
\alias{run_full_binary_pipeline}
\title{run full binary pipeline}
\usage{
run_full_binary_pipeline(file, window_size)
}
\arguments{
\item{file}{input file}

\item{window_size}{window size}
}
\value{
a list containing:
\itemize{
\item a dataframe with a combined acceleration vector and applied threshold filter
\item a plot of the acceleration vector and filtered vector
}
}
\description{
run full binary pipeline
}
