% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{run_full_threshold_pipeline}
\alias{run_full_threshold_pipeline}
\title{run full threshold pipeline}
\usage{
run_full_threshold_pipeline(file, window_size, threshold)
}
\arguments{
\item{file}{input file}

\item{window_size}{window size}

\item{threshold}{threshold to use}
}
\value{
a list containing:
\itemize{
\item a dataframe with a combined acceleration vector and applied threshold filter
\item a plot of the acceleration vector and filtered vector
}
}
\description{
run full threshold pipeline
}
\examples{
csvfile <- system.file("extdata", "Example_window.csv", package = "slidingwindow")
slidingwindow::run_full_threshold_pipeline(csvfile,1000,0.5)

}
