% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{slide_window_threshold}
\alias{slide_window_threshold}
\title{sliding window threshold}
\usage{
slide_window_threshold(df, col = "acc_vector", window_size, threshold)
}
\arguments{
\item{df}{a dataframe}

\item{col}{column containing the acceleration vector}

\item{window_size}{window size to use}

\item{threshold}{threshold to use}
}
\value{
a dataframe with a column \verb{window result} containing the filtered
acceleration vector
}
\description{
Applies sliding window filter using threshold method.
}
