% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sliding_window_file.R
\name{threshold_filter}
\alias{threshold_filter}
\title{Threshold filter}
\usage{
threshold_filter(window_data, threshold)
}
\arguments{
\item{window_data}{The data to be parsed through the function}

\item{threshold}{The threshold value to compare against the data}
}
\value{
Iterates through a window to check if values are greater than a provided threshold, if so the value of highest difference to the reference value is returned
}
\description{
Threshold filter
}
