% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{threshold_window}
\alias{threshold_window}
\title{threshold window}
\usage{
threshold_window(window_vec, threshold)
}
\arguments{
\item{window_vec}{a vector containing a window}

\item{threshold}{threshold to use}
}
\value{
a vector with applied threshold
}
\description{
Takes a window and returns the acceleration difference if it exceeds
given threshold.
}
