% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slidingwindowfunction.R
\name{window_binary}
\alias{window_binary}
\title{Assign binary +1 or -1}
\usage{
window_binary(window)
}
\arguments{
\item{window}{window value}
}
\value{
+1 or -1 given calculation
}
\description{
Assign binary +1 or -1
}
