% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Window function.R
\name{window_function}
\alias{window_function}
\title{Sliding Window Function}
\usage{
window_function(vector, window_size, threshold)
}
\arguments{
\item{vector}{Input vector (acceleration data).}

\item{window_size}{Size of the window.}

\item{threshold}{Threshold value for acceleration change.}
}
\value{
A list with mean and binary values.
}
\description{
Sliding Window Function
}
