% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slidingwindowfunction.R
\name{window_loop}
\alias{window_loop}
\title{Apply iteration on the functions}
\usage{
window_loop(data, window_size, window_function, ...)
}
\arguments{
\item{data}{data frame}

\item{window_size}{size of the sliding window}

\item{window_function}{functions created above}
}
\value{
a data frame
}
\description{
Apply iteration on the functions
}
