% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slidingwindowfunction.R
\name{window_mean}
\alias{window_mean}
\title{calculate mean of the window}
\usage{
window_mean(window)
}
\arguments{
\item{window}{window value}
}
\value{
mean of the window
}
\description{
calculate mean of the window
}
